Refactors GlobalSnackbar and AppError components into kolibri-common package
