+++ packages/kolibri-common/components/AppError/TechnicalTextBlock.vue
