+++ packages/kolibri-common/components/AppError/app-error-bird.png
