--- packages/kolibri-common/index.js
@@ -1,6 +1,0 @@
-'use strict';
-
-import AppError from '../../kolibri/core/assets/src/views/AppError';
-import GlobalSnackbar from '../../kolibri/core/assets/src/views/GlobalSnackbar';
-
-export { AppError, GlobalSnackbar };
